requirements.txt
# ChildFocus Backend — requirements.txt
# Python 3.13 | Windows 11 | Updated March 9, 2026
#
# Install:  pip install -r requirements.txt
# Upgrade:  pip install -r requirements.txt --upgrade
#
# SYSTEM REQUIREMENT (install separately — NOT via pip):
#   ffmpeg >= 8.0.1  →  https://www.gyan.dev/ffmpeg/builds/
#   Verify:  ffmpeg -version
#   Node.js  v24     →  https://nodejs.org/
#   Verify:  node --version

# ── Web Framework ─────────────────────────────────────────────────────────────
Flask==3.1.3
Werkzeug==3.1.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
blinker==1.9.0
click==8.3.1

# ── Database ──────────────────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.48
alembic==1.18.4
Mako==1.3.10
greenlet==3.3.2

# ── YouTube / Video Download ──────────────────────────────────────────────────
yt-dlp>=2026.2.21

# ── Computer Vision ───────────────────────────────────────────────────────────
# IMPORTANT: numpy must stay < 2.3 to avoid opencv conflict
numpy>=2.0,<2.3
opencv-python-headless==4.13.0.92

# ── Image Processing ──────────────────────────────────────────────────────────
Pillow>=11.0.0

# ── Audio Analysis ────────────────────────────────────────────────────────────
librosa==0.11.0
soundfile==0.13.1
audioread==3.1.0
audioop-lts==0.2.2
soxr==1.0.0
llvmlite==0.46.0
numba==0.64.0
pooch==1.9.0
lazy_loader==0.4
msgpack==1.1.2
decorator==5.2.1

# ── Scientific Computing ──────────────────────────────────────────────────────
scipy==1.17.1

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn==1.8.0
joblib==1.5.3
threadpoolctl==3.6.0

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==3.0.1
python-dateutil==2.9.0.post0
tzdata==2025.3
six==1.17.0

# ── HTTP / Networking ─────────────────────────────────────────────────────────
requests==2.32.5
urllib3==2.6.3
certifi>=2026.2.25
charset-normalizer==3.4.4
idna==3.11

# ── Environment & Utilities ───────────────────────────────────────────────────
python-dotenv>=1.2.2
cffi==2.0.0
pycparser==3.0
packaging==26.0
platformdirs==4.9.2
colorama==0.4.6
typing_extensions==4.15.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==9.0.2
pluggy==1.6.0
iniconfig==2.3.0
pygments==2.19.2

# ── Standard Library Backports (Python 3.13) ──────────────────────────────────
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
